# moviedb/docs/swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: urn:tos
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
host: localhost:9000
basePath: /
tags:
  - name: Controller
    description: Controller
paths:
  /authorization:
    get:
      tags:
        - Controller
      summary: Форма авторизации
      operationId: authorizationUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ModelAndView'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    head:
      tags:
        - Controller
      summary: Форма авторизации
      operationId: authorizationUsingHEAD
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ModelAndView'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      tags:
        - Controller
      summary: Форма авторизации
      operationId: authorizationUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ModelAndView'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - Controller
      summary: Форма авторизации
      operationId: authorizationUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ModelAndView'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - Controller
      summary: Форма авторизации
      operationId: authorizationUsingDELETE
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ModelAndView'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    options:
      tags:
        - Controller
      summary: Форма авторизации
      operationId: authorizationUsingOPTIONS
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ModelAndView'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    patch:
      tags:
        - Controller
      summary: Форма авторизации
      operationId: authorizationUsingPATCH
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ModelAndView'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /logins:
    post:
      tags:
        - Controller
      summary: Вход пользователя
      operationId: loginUserUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: login
          in: query
          description: login
          required: true
          type: string
        - name: password
          in: query
          description: password
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /message/movie:
    post:
      tags:
        - Controller
      summary: Получить список фильмов
      operationId: messageMovieUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: count
          in: query
          description: count
          required: true
          type: integer
          format: int32
        - name: genres
          in: query
          description: genres
          required: false
          type: string
        - name: key
          in: query
          description: key
          required: true
          type: string
        - name: offset
          in: query
          description: offset
          required: true
          type: integer
          format: int32
        - name: type
          in: query
          description: type
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Movie'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /message/movie/all_comments:
    post:
      tags:
        - Controller
      summary: Получить все комментарии к фильму
      operationId: overallCommentsUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: movieID
          in: query
          description: movieID
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/CommentMovies'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /message/movie/all_rating:
    get:
      tags:
        - Controller
      summary: Получить общий рейтинг всех фильмов
      operationId: overallRatingUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/MovieRating'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /message/movie/comment:
    post:
      tags:
        - Controller
      summary: Добавить комментарий к фильму
      operationId: commentMoviesUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: comment
          in: query
          description: comment
          required: true
          type: string
        - name: key
          in: query
          description: key
          required: true
          type: string
        - name: movieID
          in: query
          description: movieID
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /message/movie/dislike:
    post:
      tags:
        - Controller
      summary: Удалить фильм из избранного
      operationId: dislikeMoviesUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: key
          in: query
          description: key
          required: true
          type: string
        - name: movieID
          in: query
          description: movieID
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Movie'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /message/movie/favorites:
    post:
      tags:
        - Controller
      summary: Получить список избранных фильмов
      operationId: favoriteMoviesUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: key
          in: query
          description: key
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Movie'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /message/movie/inform:
    post:
      tags:
        - Controller
      summary: Получить список фильмов, добавленных и удаленных из избранного
      operationId: likeAndDislikeMoviesUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: key
          in: query
          description: key
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Movie'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /message/movie/like:
    post:
      tags:
        - Controller
      summary: Добавить фильм в избранное
      operationId: likeMoviesUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: key
          in: query
          description: key
          required: true
          type: string
        - name: movieID
          in: query
          description: movieID
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Movie'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /message/movie/rating:
    post:
      tags:
        - Controller
      summary: Установить рейтинг фильму
      operationId: ratingUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: key
          in: query
          description: key
          required: true
          type: string
        - name: movieID
          in: query
          description: movieID
          required: true
          type: integer
          format: int32
        - name: rating
          in: query
          description: rating
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: integer
            format: int32
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /message/movie/user_rating:
    post:
      tags:
        - Controller
      summary: Получить рейтинг пользователя
      operationId: userRatingUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: key
          in: query
          description: key
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/MovieRating'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /message/user:
    post:
      tags:
        - Controller
      summary: Отправить сообщение пользователю
      operationId: messageUserUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: key
          in: query
          description: key
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /ping:
    post:
      tags:
        - Controller
      summary: Проверить активность пользователя
      operationId: pingUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: key
          in: query
          description: key
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /register:
    post:
      tags:
        - Controller
      summary: Регистрация нового пользователя
      operationId: registerUserUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: login
          in: query
          description: login
          required: true
          type: string
        - name: name
          in: query
          description: name
          required: true
          type: string
        - name: password
          in: query
          description: password
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /search:
    post:
      tags:
        - Controller
      summary: Поиск фильмов
      operationId: searchMoviesUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: name
          in: query
          description: name
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Movie'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
definitions:
  CommentMovies:
    type: object
    properties:
      comment:
        type: string
      movieId:
        type: integer
        format: int32
      userLogin:
        type: string
    title: CommentMovies
  ModelAndView:
    type: object
    properties:
      empty:
        type: boolean
      model:
        type: object
      modelMap:
        type: object
        additionalProperties:
          type: object
      reference:
        type: boolean
      status:
        type: string
        enum:
          - ACCEPTED
          - ALREADY_REPORTED
          - BAD_GATEWAY
          - BAD_REQUEST
          - BANDWIDTH_LIMIT_EXCEEDED
          - CHECKPOINT
          - CONFLICT
          - CONTINUE
          - CREATED
          - DESTINATION_LOCKED
          - EXPECTATION_FAILED
          - FAILED_DEPENDENCY
          - FORBIDDEN
          - FOUND
          - GATEWAY_TIMEOUT
          - GONE
          - HTTP_VERSION_NOT_SUPPORTED
          - IM_USED
          - INSUFFICIENT_SPACE_ON_RESOURCE
          - INSUFFICIENT_STORAGE
          - INTERNAL_SERVER_ERROR
          - I_AM_A_TEAPOT
          - LENGTH_REQUIRED
          - LOCKED
          - LOOP_DETECTED
          - METHOD_FAILURE
          - METHOD_NOT_ALLOWED
          - MOVED_PERMANENTLY
          - MOVED_TEMPORARILY
          - MULTIPLE_CHOICES
          - MULTI_STATUS
          - NETWORK_AUTHENTICATION_REQUIRED
          - NON_AUTHORITATIVE_INFORMATION
          - NOT_ACCEPTABLE
          - NOT_EXTENDED
          - NOT_FOUND
          - NOT_IMPLEMENTED
          - NOT_MODIFIED
          - NO_CONTENT
          - OK
          - PARTIAL_CONTENT
          - PAYLOAD_TOO_LARGE
          - PAYMENT_REQUIRED
          - PERMANENT_REDIRECT
          - PRECONDITION_FAILED
          - PRECONDITION_REQUIRED
          - PROCESSING
          - PROXY_AUTHENTICATION_REQUIRED
          - REQUESTED_RANGE_NOT_SATISFIABLE
          - REQUEST_ENTITY_TOO_LARGE
          - REQUEST_HEADER_FIELDS_TOO_LARGE
          - REQUEST_TIMEOUT
          - REQUEST_URI_TOO_LONG
          - RESET_CONTENT
          - SEE_OTHER
          - SERVICE_UNAVAILABLE
          - SWITCHING_PROTOCOLS
          - TEMPORARY_REDIRECT
          - TOO_EARLY
          - TOO_MANY_REQUESTS
          - UNAUTHORIZED
          - UNAVAILABLE_FOR_LEGAL_REASONS
          - UNPROCESSABLE_ENTITY
          - UNSUPPORTED_MEDIA_TYPE
          - UPGRADE_REQUIRED
          - URI_TOO_LONG
          - USE_PROXY
          - VARIANT_ALSO_NEGOTIATES
      view:
        $ref: '#/definitions/View'
      viewName:
        type: string
    title: ModelAndView
  Movie:
    type: object
    properties:
      ageRestriction:
        type: string
      budget:
        type: string
      countryRu:
        type: string
      createdAt:
        type: string
      description:
        type: string
      genres:
        type: string
      id:
        type: integer
        format: int32
      name:
        type: string
      originalName:
        type: string
      persons:
        type: string
      poster:
        type: string
      slogan:
        type: string
      time:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      year:
        type: string
    title: Movie
  MovieRating:
    type: object
    properties:
      id:
        type: integer
        format: int32
      name:
        type: string
      rating:
        type: string
    title: MovieRating
  View:
    type: object
    properties:
      contentType:
        type: string
    title: View